Add actual tests for clean

+++ workspace/commands/clean.py
@@ -64,9 +64,10 @@ class Clean(AbstractCommand):
                 for repo in repos(path):
                     name = product_name(repo)
                     modified_time = os.stat(repo).st_mtime
-                    if name not in keep_products or keep_time and modified_time < keep_time:
+                    if keep_products and name not in keep_products or keep_time and modified_time < keep_time:
                         status = stat_repo(repo, return_output=True)
-                        if (not status or 'nothing to commit' in status and 'working directory clean' in status and
+                        if (not status or 'nothing to commit' in status and
+                                ('working directory clean' in status or 'working tree clean' in status) and
                                 len(all_branches(repo)) <= 1):
                             shutil.rmtree(repo)
                             removed_products.append(name)
